requirements.txt
Pillow~=10.0.0
setuptools~=52.0.0